refactor: types.ts를 _components에서 해당 page 디렉토리로 이동
