+++ src/app/library/detail/[id]/_components/CreateContentModal.tsx
@@ -1,7 +1,7 @@
 import { Modal } from '@/components/common/Modal/Modal';
 import Button from '@/components/common/Button/Button';
 import TextEditor from '@/components/common/TextEditor/TextEditor';
-import { CreateContentModalProps } from '@/app/library/detail/[id]/_components/type';
+import { CreateContentModalProps } from '@/app/library/detail/[id]/type';
 import { useRef, FormEvent } from 'react';
 import usePostContent from '@/hooks/api/supabase/stories/usePostContent';
 import { useStoryModal } from '@/providers/StoryWriteOrApproveModalProviders';
@@ -96,7 +96,7 @@ const CreateContentModal = ({
       className="flex h-full w-full max-w-232 flex-col justify-center md:block md:h-auto md:w-fit"
     >
       <form onSubmit={handlePostStoryContent} className="mb-1 pt-12">
-        <div className="flex justify-between items-end mb-3 md:px-2">
+        <div className="mb-3 flex items-end justify-between md:px-2">
           <h2 className="text-xl md:text-xl">이어질 이야기를 작성해주세요</h2>
           <p className="text-end text-gray-500">
             글자 수 제한 : {maxContentLength}
