+++ src/app/library/detail/[id]/type.ts
